lanchain/requirements.txt
langchain
openai
pypdf
PyPDF2
tiktoken
pinecone-client
optimum
python-dotenv
cchardet
chromadb
Flask
ctransformers
bitsandbytes
scipy